+++ scripts/voucherHubUtils.ts
@@ -2,8 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { ContractFactory } from 'ethers';
-import { ethers, upgrades } from 'hardhat';
-import { VoucherHub } from '../typechain-types';
+import hre, { ethers, upgrades } from 'hardhat';
+import { getDeploymentConfig } from '../deploy/deploy';
+import { VoucherHub, VoucherProxy__factory } from '../typechain-types';
 
 export async function deployHub(
     admin: string,
@@ -39,58 +40,56 @@ export async function upgradeProxy(
     const voucherHubUpgrade = contractUpgrade as VoucherHub;
     await voucherHubUpgrade.waitForDeployment();
     const voucherBeaconAddress = await voucherHubUpgrade.getVoucherBeacon();
-    const expectedHashes = await getVoucherProxyCreationCodeHash(voucherBeaconAddress);
-    const actualHash = await getVoucherProxyCreationCodeHashFromStorage(voucherHubAddress);
-    if (!expectedHashes.includes(actualHash)) {
+    const expectedHash = await getExpectedVoucherProxyCodeHash(voucherBeaconAddress);
+    const actualHash = await voucherHubUpgrade.getVoucherProxyCodeHash();
+    if (actualHash != expectedHash) {
         throw new Error(
             'Voucher proxy code hash in the new VoucherHub implementation does not match the real hash ' +
-                `[actual: ${actualHash}, expected:${expectedHashes}]`,
+                `[actual: ${actualHash}, expected:${expectedHash}]`,
         );
     }
     return voucherHubUpgrade;
 }
 
 /**
- * Read the value of the VoucherProxy creationCode hash from the storage of the
- * VoucherHub contract.
- * @param voucherHubAddress
+ * Get the expected VoucherProxy creationCode hash to prevent botched upgrades.
  * @returns value of the hash
  */
-export async function getVoucherProxyCreationCodeHashFromStorage(voucherHubAddress: string) {
-    // See contracts/VoucherHub.sol
-    const voucherHubStorageSlot = BigInt(
-        '0xfff04942078b704e33df5cf14e409bc5d715ca54e60a675b011b759db89ef800',
-    );
-    const codeHashSlot = voucherHubStorageSlot + 2n;
-    return await ethers.provider.getStorage(voucherHubAddress, codeHashSlot);
-}
-
-/**
- * Get a hardcoded value of the VoucherProxy creationCode hash to prevent
- * botched upgrades.
- *
- * @param voucherBeaconAddress
- * @returns value of the hash
- */
-export async function getVoucherProxyCreationCodeHash(voucherBeaconAddress: string) {
-    const factory = await ethers.getContractFactory('VoucherProxy');
-    const tx = await factory.getDeployTransaction(voucherBeaconAddress);
-    /**
-     * tx.data is the same as Solidity value of:
-     * ```
-     * abi.encodePacked(
-     *     type(VoucherProxy).creationCode, // bytecode
-     *     abi.encode($._voucherBeacon) // constructor args
-     * )
-     * ```
-     */
-    return ethers.keccak256(tx.data);
-
-    // TODO comment the implementation and return a hardcoded hash.
-    // For some reason, the hash is different between "hardhat test" and "hardhat coverage".
-    // '0x6b4bda6ca928b9724d26ee10eb17168dcd9c632e1905c854c10b78a05cd83398' // hardhat test
-    // '0x97822cb09c6322b4ccd1c4bb70005e4a3ce60099d392676696ce3d3d69e8949f' // hardhat coverage
-
-    // console.log(ethers.keccak256(tx.data));
-    // return '<hash>'
+export async function getExpectedVoucherProxyCodeHash(voucherBeaconAddress: string) {
+    const chainId = (await ethers.provider.getNetwork()).chainId.toString();
+    const config = await getDeploymentConfig(Number(chainId));
+    if (!config.factory || (hre as any).__SOLIDITY_COVERAGE_RUNNING) {
+        /**
+         * @dev Voucher proxy code hash is different from the production one:
+         * - when running "test" without generic factory since voucher beacon address
+         *   becomes not-deterministic (the EOA deployer nonce is involved).
+         * - when running "coverage" since the later compiles contracts with altered
+         *   solc options.
+         *
+         * The expected returned value below mimics the following Solidity piece of code:
+         * ```
+         * abi.encodePacked(
+         *     type(VoucherProxy).creationCode, // bytecode
+         *     abi.encode($._voucherBeacon) // constructor args
+         * )
+         * ```
+         */
+        return ethers.solidityPackedKeccak256(
+            ['bytes', 'bytes32'],
+            [
+                VoucherProxy__factory.bytecode, // bytecode
+                ethers.zeroPadValue(voucherBeaconAddress, 32), // constructor args
+            ],
+        );
+    } else {
+        /**
+         * @dev This is the voucher proxy code hash value expected in production,
+         * where contracts are deployed through a generic factory, hence having
+         * deterministic addresses.
+         *
+         * Note: Look very carefully before updating this value to avoid messing with
+         * existing vouchers already deployed in production.
+         */
+        return '0xebe0c40d5e931d865b61665b4e9435cc65ece10955457af1b4271e3438e4949e';
+    }
 }